Move editorConfig code to a new file
