+++ Extension/src/LanguageServer/Providers/documentRangeFormattingEditProvider.ts
@@ -4,7 +4,8 @@
  * ------------------------------------------------------------------------------------------ */
 import * as vscode from 'vscode';
 import { DefaultClient, FormatParams, FormatRangeRequest, FormatResult } from '../client';
-import { CppSettings, getEditorConfigSettings } from '../settings';
+import { getEditorConfigSettings } from '../editorConfig';
+import { CppSettings } from '../settings';
 import { makeVscodeTextEdits } from '../utils';
 
 export class DocumentRangeFormattingEditProvider implements vscode.DocumentRangeFormattingEditProvider {
